Change Span primitive to Text and use 'as' prop for tag
