+++ src/containers/PageBuilder/BlockBuilder/BlockFooter/BlockFooter.module.css
@@ -1,5 +1,3 @@
-
-
 .root {
   /* fallback option */
   word-break: break-all;
