--- src/containers/PageBuilder/Primitives/Span/index.js
@@ -1,1 +1,0 @@
-export { Span } from './Span';
