+++ src/containers/PageBuilder/Primitives/Text/index.js
@@ -1,0 +1,1 @@
+export { Text } from './Text';
