+++ src/containers/PageBuilder/SectionBuilder/SectionFooter/SectionFooter.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { arrayOf, bool, func, node, number, object, shape, string } from 'prop-types';
 import classNames from 'classnames';
-import {LinkedLogo} from '../../../../components';
+import { LinkedLogo } from '../../../../components';
 
 import Field, { hasDataInFields } from '../../Field';
 import BlockBuilder from '../../BlockBuilder';
@@ -17,19 +17,19 @@ const GRID_CONFIG = [
   { contentCss: css.contentCol2, gridCss: css.gridCol2 },
   { contentCss: css.contentCol3, gridCss: css.gridCol3 },
   { contentCss: css.contentCol4, gridCss: css.gridCol4 },
-]
+];
 
 const getIndex = numberOfColumns => numberOfColumns - 1;
 
 const getContentCss = numberOfColumns => {
   const contentConfig = GRID_CONFIG[getIndex(numberOfColumns)];
   return contentConfig ? contentConfig.contentCss : GRID_CONFIG[0].contentCss;
-}
+};
 
 const getGridCss = numberOfColumns => {
   const contentConfig = GRID_CONFIG[getIndex(numberOfColumns)];
   return contentConfig ? contentConfig.gridCss : GRID_CONFIG[0].gridCss;
-}
+};
 
 // Section component that's able to show blocks in multiple different columns (defined by "numberOfColumns" prop)
 const SectionFooter = props => {
@@ -50,7 +50,7 @@ const SectionFooter = props => {
   // E.g. { h1: { component: MyAwesomeHeader } }
   const fieldComponents = options?.fieldComponents;
   const fieldOptions = { fieldComponents };
-  console.log({ socialMediaLinks })
+  console.log({ socialMediaLinks });
 
   const showSocialMediaLinks = socialMediaLinks?.length > 0;
 
@@ -58,37 +58,33 @@ const SectionFooter = props => {
 
   return (
     <footer>
-    <SectionContainer
-      id={sectionId}
-      className={className}
-      rootClassName={rootClassName}
-      appearance={appearance}
-      options={fieldOptions}
-    >
-      <div className={css.footer}>
-        <div className={classNames(css.content, getContentCss(numberOfColumns))}>
-          <div>
-            <LinkedLogo className={css.logoLink} />
-          </div>
-        <div className={css.detailsInfo}>
-            <Field data={slogan}/>
-          </div>
-          <div className={css.socialInfo}>
-            <div className={css.icons}>
-              {showSocialMediaLinks ? (
-                <BlockBuilder blocks={socialMediaLinks} />
-              ) : null}
+      <SectionContainer
+        id={sectionId}
+        className={className}
+        rootClassName={rootClassName}
+        appearance={appearance}
+        options={fieldOptions}
+      >
+        <div className={css.footer}>
+          <div className={classNames(css.content, getContentCss(numberOfColumns))}>
+            <div>
+              <LinkedLogo className={css.logoLink} />
             </div>
-            <Field data={copyright} />
-          </div>
-          <div className={classNames(css.grid, getGridCss(numberOfColumns))}>
-            <BlockBuilder
-              blocks={blocks}
-            />
+            <div className={css.detailsInfo}>
+              <Field data={slogan} />
+            </div>
+            <div className={css.socialInfo}>
+              <div className={css.icons}>
+                {showSocialMediaLinks ? <BlockBuilder blocks={socialMediaLinks} /> : null}
+              </div>
+              <Field data={copyright} />
+            </div>
+            <div className={classNames(css.grid, getGridCss(numberOfColumns))}>
+              <BlockBuilder blocks={blocks} />
+            </div>
           </div>
         </div>
-      </div>
-    </SectionContainer>
+      </SectionContainer>
     </footer>
   );
 };
